e school page
I wrote this site with PHP, students can enter their grades, they can enter their absences, there is a course schedule, there is a section for entering teacher grades, a section for creating a course schedule, a section for entering absences. I created a parent login so that parents can log in to check student grades without the student login.